pyproject.toml
[project]
name = "agent-development-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.3.0",
    "litellm>=1.72.6.post1",
    "requests>=2.32.4",
    "yfinance>=0.2.63",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["multi_agent/test"]
python_files = "test_*.py"

